Refactor AppListScreen to use Scaffold and add TopAppBar
This commit refactors the `AppListScreen` to utilize `Scaffold` for its layout.
A new `AppListTopBar` has been introduced which includes the search bar and a dropdown menu with navigation options to "Repositories" and "Settings".

The navigation for `AppListScreen` has been updated to include these new navigation actions.
Additionally, the `startDestination` in `MainComposeActivity` has been changed from `Home` to `AppList`.